✨ Update the Bottom Sheet implementation to use Accompanist
The previous/manual implementation of Alkaa is really complex to
maintain once the navigation was made manually and with States.
The code is now updated to use the NavGraph to navigate through the
Bottom Sheets.
Also, some changes were needed once the CategoryBottomSheet received an
internal Category representation, now it receives only the category id.
This change required some updates and new states in the implementation.